Refactoring V1

+++ APIServices/Zendesk_Instagram/Services/ZendeskChannelService.php
@@ -41,28 +41,34 @@ class ZendeskChannelService {
      */
     public function getUpdates()
     {
-        try {
-            $owner_post = $this->instagram_service->getOwnerInstagram();
-            $response = $this->instagram_service->getInstagramPosts(199);
-            $posts = $response['data'];
-            //It is done to start with the oldest post, to show properly in Zendes.
-            $posts = array_reverse($posts, false);
-            $post_timestamp = $this->state;
-            $transformedMessages = [];
-            foreach ($posts as $post) {
-                if (count($transformedMessages) > 195) {
-                    break;
-                }
-                $post_id = $post['id'];
-                $post_timestamp = date("c", strtotime($post['timestamp']));
-                if ($this->expire($post_timestamp)) {
-                    $this->instagram_service->removePost($post_id);
-                    continue;
-                }
-                if ($post_timestamp > $this->state['last_post_date']) {
-                    array_push($transformedMessages, $this->getUpdatesPosts($owner_post,$post));
-                }
-                $response = $this->instagram_service->getInstagramCommentsFromPost($post_id);
+        $transformedMessages = [];
+        $post_timestamp = $this->state;
+        $owner_post = $this->instagram_service->getOwner();
+        if ($owner_post->isError()){
+            return $this->getResponsePull($transformedMessages,$post_timestamp);
+        }
+        $response_post = $this->instagram_service->getPosts(199);
+        if ($response_post->isError()){
+            return $this->getResponsePull($transformedMessages,$post_timestamp);
+        }
+        $posts = $response_post['data'];
+        //It is done to start with the oldest post, to show properly in Zendes.
+        $posts = array_reverse($posts, false);
+        foreach ($posts as $post) {
+            if (count($transformedMessages) > 195) {
+                break;
+            }
+            $post_id = $post['id'];
+            $post_timestamp = date("c", strtotime($post['timestamp']));
+            if ($this->expire($post_timestamp)) {
+                $this->instagram_service->removePost($post_id);
+                continue;
+            }
+            if ($post_timestamp > $this->state['last_post_date']) {
+                array_push($transformedMessages, $this->getUpdatesPosts($owner_post,$post));
+            }
+            $response = $this->instagram_service->getCommentsFromPost($post_id);
+            if (!$response->isError()){
                 $comments = $response['data'];
                 //It is done to start with the oldest post, to show properly in Zendes.
                 $comments = array_reverse($comments, false);
@@ -74,32 +80,39 @@ class ZendeskChannelService {
                     $comment_timestamp = date("c", strtotime($comment['timestamp']));
                     $comment_timestamp = new Carbon($comment_timestamp);
                     $comment_track = $this->instagram_service->commentTrack($post_id,$comment_timestamp);
-                    $last_comment_date = $comment_track->last_comment_date;
-                    if ($comment_timestamp >= $last_comment_date) {
-                        array_push($transformedMessages, $this->getUpdatesComments($owner_post, $post_id, $comment));
-                        $last_comment_date = $comment_timestamp;
+                    if ($comment_track->isSuccess()){
+                        $last_comment_date = $comment_track->last_comment_date;
+                        if ($comment_timestamp >= $last_comment_date) {
+                            array_push($transformedMessages, $this->getUpdatesComments($owner_post, $post_id, $comment));
+                            $last_comment_date = $comment_timestamp;
+                        }
                     }
                 }
-               $this->instagram_service->updatePost($post_id, $last_comment_date);
-
+                //To update the date of the last comment
+                $this->instagram_service->updatePost($post_id, $last_comment_date);
             }
-            //To Zendesk Pull
-            $response = [
-                'external_resources' => $transformedMessages,
-                'state' => json_encode(['last_post_date' => sprintf('%s', $post_timestamp)])
-            ];
-            Log::debug($response);
-            return $response;
-        } catch (\Exception $exception) {
-//            return [
-//                'external_resources' => [],
-//                'state' => '{}',
-//                'metadata_needs_update' => true
-//            ];
-            return ['Message Error: ' + $exception->getMessage()];
+
         }
+        return $this->getResponsePull($transformedMessages,$post_timestamp);
     }
 
+    /**
+     * @param $transformedMessages
+     * @param $post_timestamp
+     * @return array
+     */
+    private function getResponsePull($transformedMessages,$post_timestamp){
+        return [
+            'external_resources' => $transformedMessages,
+            'state' => json_encode(['last_post_date' => sprintf('%s', $post_timestamp)])
+        ];
+    }
+
+    /**
+     * @param $owner_post
+     * @param $post
+     * @return array
+     */
     private function getUpdatesPosts($owner_post, $post)
     {
         /** @var PostFormatter $formatter */
@@ -111,6 +124,12 @@ class ZendeskChannelService {
         return $formatter->getTransformedMessage();
     }
 
+    /**
+     * @param $owner_post
+     * @param $post_id
+     * @param $comment
+     * @return array
+     */
     private function getUpdatesComments($owner_post, $post_id, $comment)
     {
         /** @var CommentFormatter $formatter */
@@ -124,6 +143,10 @@ class ZendeskChannelService {
         return $formatter->getTransformedMessage();
     }
 
+    /**
+     * @param $date
+     * @return bool
+     */
     private function expire($date)
     {
         $date = new Carbon($date);
